saw-script interpreter: move IsValue and FromValue to ValueOps.hs
This moves them out of saw-central so they're no longer exposed.

Note that ValueOps is not necessarily the right place. There'll be a
reorg in saw-script later; for now the priority is to move stuff that
doesn't belong in saw-central out.

Update the test1646 reference output that tickles a crash in
FromValue; it has a Haskell-source file and line number in it.
