+++ intTests/test1646/test2.log.good
@@ -2,5 +2,5 @@
  x=a
  fromValue String
 CallStack (from HasCallStack):
-  error, called at saw-central/src/SAWCentral/Value.hs:1247:19 in saw-1.3.0.99-inplace-saw-central:SAWCentral.Value
+  error, called at saw-script/src/SAWScript/ValueOps.hs:619:19 in saw-1.3.0.99-inplace-saw-script:SAWScript.ValueOps
 FAILED
